ia-backend/requirements.txt
fastapi
uvicorn
python-multipart
azure-storage-blob
requests
pyodbc